refactor: 카드 타입 선언을 api 디렉토리로 옮겨 여러 페이지에서 공통으로 사용할 수 있도록 함
