+++ src/api/card/types.ts
@@ -1,0 +1,5 @@
+export type CardType =
+  | 'GREETING'
+  | 'CELEBRATION'
+  | 'ENCOURAGEMENT'
+  | 'GRATITUDE';
