+++ src/components/pages/my-card/index.tsx
@@ -2,7 +2,8 @@ import { useEffect, useState } from 'react';
 import { HStack, Heading } from '@chakra-ui/react';
 import ReactPaginate from 'react-paginate';
 import { ProfileInfo } from '@/components/common/profile-info/types';
-import { CardDirection, CardViewType, CardInfo, CardType } from './types';
+import { CardType } from '@/api/card/types';
+import { CardDirection, CardViewType, CardInfo } from './types';
 import { ProfileCardSection, CardListSection, CardListFooter } from './styles';
 import GNBLayout from '@/layouts/gnb-layout';
 import ProfileCard from '@/components/pages/my-card/profile-card/index';
